FLAS-25 Fit dialog form
