--- assets/components/dialog.tsx
@@ -1,66 +1,0 @@
-import * as React from 'react';
-import Button from '@mui/material/Button';
-import TextField from '@mui/material/TextField';
-import Dialog from '@mui/material/Dialog';
-import DialogActions from '@mui/material/DialogActions';
-import DialogContent from '@mui/material/DialogContent';
-import DialogContentText from '@mui/material/DialogContentText';
-import DialogTitle from '@mui/material/DialogTitle';
-
-interface FormDialogProps {
-    /**
-     * TODO
-     *  1. Ustawić typ opena
-     *  2. Zmienić wnętrze dialogu, żeby można było jako child oraz tekst wyświetlany w oknie
-     */
-
-    open: any;
-    handleClose: () => void;
-}
-
-export default function FormDialog({open, handleClose}: FormDialogProps): React.ReactElement {
-    return (
-        <React.Fragment>
-            <Dialog
-                open={open}
-                onClose={handleClose}
-                slotProps={{
-                    paper: {
-                        component: 'form',
-                        onSubmit: (event: React.FormEvent<HTMLFormElement>) => {
-                            event.preventDefault();
-                            const formData = new FormData(event.currentTarget);
-                            const formJson = Object.fromEntries((formData as any).entries());
-                            const email = formJson.email;
-                            console.log(email);
-                            handleClose();
-                        },
-                    },
-                }}
-            >
-                <DialogTitle>Subscribe</DialogTitle>
-                <DialogContent>
-                    <DialogContentText>
-                        To subscribe to this website, please enter your email address here. We
-                        will send updates occasionally.
-                    </DialogContentText>
-                    <TextField
-                        autoFocus
-                        required
-                        margin="dense"
-                        id="name"
-                        name="email"
-                        label="Email Address"
-                        type="email"
-                        fullWidth
-                        variant="standard"
-                    />
-                </DialogContent>
-                <DialogActions>
-                    <Button onClick={handleClose}>Cancel</Button>
-                    <Button type="submit">Subscribe</Button>
-                </DialogActions>
-            </Dialog>
-        </React.Fragment>
-    );
-}
